requirements.txt
jupyter==1.0.0
numpy==1.19.2
pandas==1.1.1
matplotlib==3.3.0
tqdm==4.48.0
pyserial==3.5
easydict==1.9
nvidia-ml-py==11.450.51